Fixed a bug with the overwrite flag and MPI. Old files were being deleted from all MPI processes, which means that some non-root process could be deleting some of the files newly created by the root process.

+++ include/CosmoInterface/measurements/measurementsIO/filesmanager.h
@@ -58,7 +58,7 @@ namespace TempLat
     {
       const bool exists = std::filesystem::exists(fn);
       if (mAppendMode) return !exists;
-      if (exists) {
+      if (exists && mToolbox->amIRoot()) {
         if (mOverwriteMode) std::filesystem::remove(fn);
         else throw(FileAlreadyExistsError(
             "Refusing to overwrite existing output file \"" + fn +
